created some nonsense

+++ db/create_db.sql
@@ -1,5 +1,5 @@
 create table users (
-    id integer primary key auto_increment,
+    id bigint unsigned primary key auto_increment,
     username varchar(50) unique not null,
     sexual_orientation varchar(20) not null,
     passwd varchar(50) not null,
@@ -8,17 +8,18 @@ create table users (
     race varchar(10)
 );
 
+
 create table pictures (
-    id integer primary key auto_increment,
-    owner_id integer not null,
+    id bigint unsigned primary key auto_increment,
+    owner_id bigint unsigned not null,
     path_to_pic varchar(100) not null,
     foreign key(owner_id) references users(id) on delete cascade
 );
 
 create table ratings (
-    id integer primary key auto_increment,
-    rater_id integer,
-    rated_id integer,
+    id bigint unsigned primary key auto_increment,
+    rater_id bigint unsigned,
+    rated_id bigint unsigned,
     mark integer,
     foreign key (rater_id) references users(id) on delete set null,
     foreign key (rated_id) references users(id) on delete set null
